requirements.txt
vosk>=0.3.45
sounddevice>=0.4.6
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0
